requirements.txt
pypdf
python-docx
pandas
openpyxl
tqdm
chromadb
sentence-transformers
rank-bm25
ollama